refactor: keytree
